requirements.txt
selenium>=4.15.0
Pillow>=10.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
openai>=0.28.0
requests>=2.31.0